Split order and lead API schema modules
